Add provisioning interface

+++ db/migrate/20140910153654_move_host_nics_to_interfaces.rb
@@ -2,7 +2,7 @@ class FakeNic < ActiveRecord::Base
   self.table_name = 'nics'
 
   attr_accessible :host_id, :name, :mac, :ip, :subnet_id, :domain_id, :identifier,
-                  :virtual, :primary, :managed
+                  :virtual, :primary, :provision, :managed
 
   def type
     Nic::Managed
@@ -12,7 +12,7 @@ end
 class FakeHost < ActiveRecord::Base
   self.table_name = 'hosts'
 
-  attr_accessible :mac, :ip, :subnet_id, :domain_id, :primary_interface
+  attr_accessible :mac, :ip, :subnet_id, :domain_id, :primary_interface, :provision_interface
 
   def type
     Host::Managed
@@ -22,8 +22,9 @@ end
 class MoveHostNicsToInterfaces < ActiveRecord::Migration
   def up
     add_column :nics, :primary, :boolean, :default => false
+    add_column :nics, :provision, :boolean, :default => false
 
-    say "Migrating Host interfaces to Nic primaries"
+    say "Migrating Host interfaces to standalone Interfaces"
     Host::Managed.all.each do |host|
       next unless host.managed?
       nic = FakeNic.new
@@ -37,7 +38,8 @@ class MoveHostNicsToInterfaces < ActiveRecord::Migration
       nic.identifier = host.primary_interface || "eth0"
       nic.managed = true
       nic.primary = true
-      nic.save(:validate => false)
+      nic.provision = true
+      nic.save!
       say "  Migrated #{nic.name}-#{nic.identifier} to nics"
     end
 
@@ -55,19 +57,23 @@ class MoveHostNicsToInterfaces < ActiveRecord::Migration
     add_column :hosts, :subnet_id, :integer
     add_column :hosts, :domain_id, :integer
 
-    say "Migrating Nic primaries to Host interfaces"
+    say "Migrating Interfaces to Host interfaces"
+    FakeHost.reset_column_information
     FakeHost.all.each do |host|
       next unless host.managed?
-      nic = FakeNic.where(:host_id => host.id).where(:primary => true).first
+      raise ActiveRecord::IrreversibleMigration if FakeNic.where(:primary => true, :provision => false).any?
+      raise ActiveRecord::IrreversibleMigration if FakeNic.where(:primary => false, :provision => true).any?
+      nic = FakeNic.where(:host_id => host.id).where(:primary => true, :provision => true).first
       host.mac = nic.mac
       host.ip = nic.ip
       host.subnet_id = nic.subnet_id
       host.domain_id = nic.domain_id
       host.primary_interface = nic.identifier
-      host.save(:validate => false)
+      host.save!
       nic.destroy!
     end
 
     remove_column :nics, :primary
+    remove_column :nics, :provision
   end
 end
